# Optimise conservative draw quad computations